# Vectorize JitteredResize