--- a/keras_cv/layers/preprocessing/jittered_resize.py
+++ b/keras_cv/layers/preprocessing/jittered_resize.py
@@ -1,4 +1,4 @@
-# Copyright 2022 The KerasCV Authors
+# Copyright 2023 The KerasCV Authors
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -19,14 +19,18 @@
 import tensorflow as tf
 from tensorflow import keras
 
-import keras_cv
-from keras_cv.layers.preprocessing.base_image_augmentation_layer import (
-    BaseImageAugmentationLayer,
+from keras_cv import bounding_box
+from keras_cv.layers.preprocessing.vectorized_base_image_augmentation_layer import (  # noqa: E501
+    VectorizedBaseImageAugmentationLayer,
 )
+from keras_cv.utils import preprocessing as preprocessing_utils
+
+H_AXIS = -3
+W_AXIS = -2
 
 
 @keras.utils.register_keras_serializable(package="keras_cv")
-class JitteredResize(BaseImageAugmentationLayer):
+class JitteredResize(VectorizedBaseImageAugmentationLayer):
     """JitteredResize implements resize with scale distortion.
 
     JitteredResize takes a three-step approach to size-distortion based image
@@ -46,7 +50,26 @@ class JitteredResize(BaseImageAugmentationLayer):
     are translated and scaled according to the random scaling and random
     cropping.
 
+    Args:
+        target_size: A tuple representing the output size of images.
+        scale_factor: A tuple of two floats or a `keras_cv.FactorSampler`. For
+            each augmented image a value is sampled from the provided range.
+            This factor is used to scale the input image.
+            To replicate the results of the MaskRCNN paper pass `(0.8, 1.25)`.
+        crop_size: (Optional) the size of the image to crop from the scaled
+            image, defaults to `target_size` when not provided.
+        bounding_box_format: The format of bounding boxes of input boxes.
+            Refer to
+            https://github.com/keras-team/keras-cv/blob/master/keras_cv/bounding_box/converters.py
+            for more details on supported bounding box formats.
+        interpolation: String, the interpolation method, defaults to
+            `"bilinear"`. Supports `"bilinear"`, `"nearest"`, `"bicubic"`,
+            `"area"`, `"lanczos3"`, `"lanczos5"`, `"gaussian"`,
+            `"mitchellcubic"`.
+        seed: (Optional) integer to use as the random seed.
+
     Usage:
+
     ```python
     train_ds = load_object_detection_dataset()
     jittered_resize = layers.JitteredResize(
@@ -72,24 +95,6 @@ class JitteredResize(BaseImageAugmentationLayer):
     )
     # images now are (640, 640, 3), but they were resized from a 250x250 crop.
     ```
-
-    Args:
-        target_size: A tuple representing the output size of images.
-        scale_factor: A tuple of two floats or a `keras_cv.FactorSampler`. For
-            each augmented image a value is sampled from the provided range.
-            This factor is used to scale the input image.
-            To replicate the results of the MaskRCNN paper pass `(0.8, 1.25)`.
-        crop_size: (Optional) the size of the image to crop from the scaled
-            image, defaults to `target_size` when not provided.
-        bounding_box_format: The format of bounding boxes of input boxes.
-            Refer to
-            https://github.com/keras-team/keras-cv/blob/master/keras_cv/bounding_box/converters.py
-            for more details on supported bounding box formats.
-        interpolation: String, the interpolation method, defaults to
-            `"bilinear"`. Supports `"bilinear"`, `"nearest"`, `"bicubic"`,
-            `"area"`, `"lanczos3"`, `"lanczos5"`, `"gaussian"`,
-            `"mitchellcubic"`.
-        seed: (Optional) integer to use as the random seed.
     """
 
     def __init__(
@@ -110,8 +115,10 @@ class JitteredResize(BaseImageAugmentationLayer):
             )
 
         crop_size = crop_size or target_size
-        self.interpolation = keras_cv.utils.get_interpolation(interpolation)
-        self.scale_factor = keras_cv.utils.parse_factor(
+        self.interpolation = preprocessing_utils.get_interpolation(
+            interpolation
+        )
+        self.scale_factor = preprocessing_utils.parse_factor(
             scale_factor,
             min_value=0.0,
             max_value=None,
@@ -122,91 +129,119 @@ class JitteredResize(BaseImageAugmentationLayer):
         self.target_size = target_size
         self.bounding_box_format = bounding_box_format
         self.seed = seed
-        self.force_output_dense_images = True
-        self.auto_vectorize = False
 
-    def get_random_transformation(self, image=None, **kwargs):
-        original_image_shape = tf.shape(image)
-        image_shape = tf.cast(original_image_shape[0:2], tf.float32)
-
-        scaled_size = tf.round(image_shape * self.scale_factor())
-        scale = tf.minimum(
-            scaled_size[0] / image_shape[0], scaled_size[1] / image_shape[1]
-        )
-
-        scaled_size = tf.round(image_shape * scale)
-        image_scale = scaled_size / image_shape
-
-        max_offset = scaled_size - self.crop_size
-        max_offset = tf.where(
-            tf.less(max_offset, 0), tf.zeros_like(max_offset), max_offset
-        )
-        offset = max_offset * tf.random.uniform([2], minval=0, maxval=1)
-        offset = tf.cast(offset, tf.int32)
-
-        return {
-            "original_size": original_image_shape,
-            "image_scale": image_scale,
-            "scaled_size": scaled_size,
-            "offset": offset,
-        }
-
-    def compute_image_signature(self, images):
-        return tf.TensorSpec(
+    def compute_ragged_image_signature(self, images):
+        ragged_spec = tf.RaggedTensorSpec(
             shape=list(self.target_size) + [images.shape[-1]],
+            ragged_rank=1,
             dtype=self.compute_dtype,
         )
+        return ragged_spec
 
-    def augment_image(self, image, transformation, **kwargs):
+    def get_random_transformation_batch(
+        self, batch_size, images=None, **kwargs
+    ):
+        heights, widths = self._get_image_shape(images)
+        image_shapes = tf.cast(
+            tf.concat((heights, widths), axis=-1), dtype=tf.float32
+        )
+
+        scaled_sizes = tf.round(
+            image_shapes * self.scale_factor(shape=(batch_size, 1))
+        )
+        scales = tf.where(
+            tf.less(
+                scaled_sizes[..., 0] / image_shapes[..., 0],
+                scaled_sizes[..., 1] / image_shapes[..., 1],
+            ),
+            scaled_sizes[..., 0] / image_shapes[..., 0],
+            scaled_sizes[..., 1] / image_shapes[..., 1],
+        )
+
+        scaled_sizes = tf.round(image_shapes * scales[..., tf.newaxis])
+        image_scales = scaled_sizes / image_shapes
+
+        max_offsets = scaled_sizes - self.crop_size
+        max_offsets = tf.where(
+            tf.less(max_offsets, 0), tf.zeros_like(max_offsets), max_offsets
+        )
+        offsets = max_offsets * self._random_generator.random_uniform(
+            shape=(batch_size, 2), minval=0, maxval=1, dtype=tf.float32
+        )
+        offsets = tf.cast(offsets, tf.int32)
+        return {
+            "image_scales": image_scales,
+            "scaled_sizes": scaled_sizes,
+            "offsets": offsets,
+        }
+
+    def augment_ragged_image(self, image, transformation, **kwargs):
+        image = tf.expand_dims(image, axis=0)
+        scaled_sizes = transformation["scaled_sizes"]
+        offsets = transformation["offsets"]
+        transformation = {
+            "scaled_sizes": tf.expand_dims(scaled_sizes, axis=0),
+            "offsets": tf.expand_dims(offsets, axis=0),
+        }
+        image = self.augment_images(
+            images=image, transformations=transformation, **kwargs
+        )
+        return tf.squeeze(image, axis=0)
+
+    def augment_images(self, images, transformations, **kwargs):
         # unpackage augmentation arguments
-        scaled_size = transformation["scaled_size"]
-        offset = transformation["offset"]
-        target_size = self.target_size
-        crop_size = self.crop_size
-
-        scaled_image = tf.image.resize(
-            image, tf.cast(scaled_size, tf.int32), method=self.interpolation
+        scaled_sizes = transformations["scaled_sizes"]
+        offsets = transformations["offsets"]
+        inputs_for_augment_images_fn = {
+            "IMAGES": images,
+            "SCALE_SIZES": scaled_sizes,
+            "OFFSETS": offsets,
+        }
+        scaled_images = tf.map_fn(
+            self._augment_images_fn,
+            inputs_for_augment_images_fn,
+            fn_output_signature=tf.float32,
         )
-        scaled_image = scaled_image[
-            offset[0] : offset[0] + crop_size[0],
-            offset[1] : offset[1] + crop_size[1],
-            :,
-        ]
-        scaled_image = tf.image.pad_to_bounding_box(
-            scaled_image, 0, 0, target_size[0], target_size[1]
-        )
-        return tf.cast(scaled_image, self.compute_dtype)
+        return tf.cast(scaled_images, self.compute_dtype)
 
-    def augment_bounding_boxes(self, bounding_boxes, transformation, **kwargs):
+    def augment_labels(self, labels, transformations, **kwargs):
+        return labels
+
+    def augment_bounding_boxes(
+        self, bounding_boxes, transformations, raw_images=None, **kwargs
+    ):
         if self.bounding_box_format is None:
             raise ValueError(
                 "Please provide a `bounding_box_format` when augmenting "
                 "bounding boxes with `JitteredResize()`."
             )
+        if isinstance(bounding_boxes["boxes"], tf.RaggedTensor):
+            bounding_boxes = bounding_box.to_dense(bounding_boxes)
         result = bounding_boxes.copy()
-        image_scale = tf.cast(transformation["image_scale"], self.compute_dtype)
-        offset = tf.cast(transformation["offset"], self.compute_dtype)
-        original_size = transformation["original_size"]
+        image_scales = tf.cast(
+            transformations["image_scales"], self.compute_dtype
+        )
+        offsets = tf.cast(transformations["offsets"], self.compute_dtype)
 
-        bounding_boxes = keras_cv.bounding_box.convert_format(
+        bounding_boxes = bounding_box.convert_format(
             bounding_boxes,
-            image_shape=original_size,
+            images=raw_images,
             source=self.bounding_box_format,
             target="yxyx",
         )
 
         # Adjusts box coordinates based on image_scale and offset.
         yxyx = bounding_boxes["boxes"]
-        yxyx *= tf.tile(tf.expand_dims(image_scale, axis=0), [1, 2])
-        yxyx -= tf.tile(tf.expand_dims(offset, axis=0), [1, 2])
+        yxyx *= tf.tile(image_scales, [1, 2])[..., tf.newaxis, :]
+        yxyx -= tf.tile(offsets, [1, 2])[..., tf.newaxis, :]
 
         result["boxes"] = yxyx
-        result = keras_cv.bounding_box.clip_to_image(
+        result = bounding_box.clip_to_image(
             result,
             image_shape=self.target_size + (3,),
             bounding_box_format="yxyx",
         )
-        result = keras_cv.bounding_box.convert_format(
+        result = bounding_box.convert_format(
             result,
             image_shape=self.target_size + (3,),
             source="yxyx",
@@ -214,8 +249,35 @@ class JitteredResize(BaseImageAugmentationLayer):
         )
         return result
 
-    def augment_label(self, label, transformation, **kwargs):
-        return label
+    def _get_image_shape(self, images):
+        if isinstance(images, tf.RaggedTensor):
+            heights = tf.reshape(images.row_lengths(), (-1, 1))
+            widths = tf.reshape(
+                tf.reduce_max(images.row_lengths(axis=2), 1), (-1, 1)
+            )
+        else:
+            batch_size = tf.shape(images)[0]
+            heights = tf.repeat(tf.shape(images)[H_AXIS], repeats=[batch_size])
+            heights = tf.reshape(heights, shape=(-1, 1))
+            widths = tf.repeat(tf.shape(images)[W_AXIS], repeats=[batch_size])
+            widths = tf.reshape(widths, shape=(-1, 1))
+        return tf.cast(heights, dtype=tf.int32), tf.cast(widths, dtype=tf.int32)
+
+    def _augment_images_fn(self, inputs):
+        image = inputs.get("IMAGES", None)
+        scaled_size = inputs.get("SCALE_SIZES", None)
+        offset = inputs.get("OFFSETS", None)
+
+        scaled_image = tf.image.resize(image, tf.cast(scaled_size, tf.int32))
+        scaled_image = scaled_image[
+            offset[0] : offset[0] + self.crop_size[0],
+            offset[1] : offset[1] + self.crop_size[1],
+            :,
+        ]
+        scaled_image = tf.image.pad_to_bounding_box(
+            scaled_image, 0, 0, self.target_size[0], self.target_size[1]
+        )
+        return scaled_image
 
     def get_config(self):
         config = super().get_config()
@@ -230,3 +292,11 @@ class JitteredResize(BaseImageAugmentationLayer):
             }
         )
         return config
+
+    @classmethod
+    def from_config(cls, config):
+        if isinstance(config["scale_factor"], dict):
+            config["scale_factor"] = keras.utils.deserialize_keras_object(
+                config["scale_factor"]
+            )
+        return cls(**config)