… benchmarks/vectorized_jittered_resize.py