# What does this PR do?

This PR might be helpful to speed up model training, such as RetinaNet:

https://github.com/keras-team/keras-cv/blob/63f7fcfe039bd573158e8f2310b48b86cf372bc6/examples/training/object_detection/pascal_voc/retina_net.py#L118-L129

## Benchmarks

WITHOUT `bounding_boxes`: vectorized JitteredResize is better than non-vectorized one.
||Mode|n=100|n=200|n=500|n=1000|
|-|-|-|-|-|-|
|OldJitteredResize|Eager|0.616s|1.238s|3.051s|6.106s|
|OldJitteredResize|Graph|0.057s|0.110s|0.277s|0.540s|
|JitteredResize|Eager|0.280s|0.548s|1.391s|2.760s|
|JitteredResize|Graph|0.046s|0.091s|0.233s|0.447s|

![comparison2](https://user-images.githubusercontent.com/20734616/230828611-60fe0d81-20ba-4c78-bbd4-9f5f4bc34b4f.png)

WITH `bounding_boxes`: vectorized JitteredResize is much better than non-vectorized one.
(Simulation: each image contains 4 boxes)

||Mode|n=100|n=200|n=500|n=1000|
|-|-|-|-|-|-|
|OldJitteredResize|Eager|2.872s|5.722s|14.322s|28.586s|
|OldJitteredResize|Graph|0.176s|0.526s|2.673s|9.624s|
|JitteredResize|Eager|0.293s|0.559s|1.376s|2.713s|
|JitteredResize|Graph|0.048s|0.095s|0.224s|0.453s|

![comparison](https://user-images.githubusercontent.com/20734616/230828631-bd231469-c2a5-4930-bc86-c913881d9f2c.png)

OldJitteredResize is too slow, therefore I have set `n` to `[100, 200, 500, 1000]`. There is no XLA mode result available due to the inability of `tf.map_fn/while_loop` to run on XLA.

## Visualization

This PR updates the working pipeline for visualization.

![output](https://user-images.githubusercontent.com/20734616/230829131-0e82b9f3-62bb-4bd8-a199-8d53a2030b02.png)

## Notes

As far as my knowledge goes, directly vectorizing `augment_images` seems impossible due to the varying `scaled_size` of each image. However this PR utilizes `tf.map_fn` to overcome it.

The drawback is that the speed improvement is not significant without `bounding_boxes`.

The output images of JitteredResize are always dense tensors.

## Before submitting
- [ ] This PR fixes a typo or improves the docs (you can dismiss the other checks if that's the case).
- [X] Did you read the [contributor guideline](https://github.com/keras-team/keras-cv/blob/master/.github/CONTRIBUTING.md),
      Pull Request section?
- [ ] Was this discussed/approved via a Github issue? Please add a link
      to it if that's the case.
- [X] Did you write any new necessary tests?
- [ ] If this adds a new model, can you run a few training steps on TPU in Colab to ensure that no XLA incompatible OP are used?

## Who can review?
@LukeWood @ianstenbit