@james77777778 2023-04-13
Hi @LukeWood @ianstenbit 

I came across a comment on https://github.com/keras-team/keras-io/pull/1314#pullrequestreview-1381235098 suggesting that JitteredResize should be renamed to RandomZoomAndCrop and replace `target_size` with `height` and `width` like other KPLs.

Should I do it in this PR?