@james77777778 2023-04-19
> LGTM -- just one small nit. With respect to your question about renaming and updating parameter structure, let's make that a separate PR. You're welcome to raise a PR for that if you'd like.
> 
> Thank you for all your contributions here!

All done.

I will submit another PR to rename and update the parameters as soon as I am available