@ianstenbit 2023-04-20
@LukeWood is going to do a training run with this change to triple-check numerics, and then I think we can merge.

Thank you @james77777778! Your work on these has been fantastic.