@james77777778 2023-04-20
@ianstenbit @LukeWood 
It would be great if we can obtain benchmark results for both the old and vectorized layers on a popular dataset such as VOC2007!