@LukeWood 2023-04-20
> @ianstenbit @LukeWood It would be great if we can obtain benchmark results for both the old and vectorized layers on a popular dataset such as VOC2007!

ideally I'll have integration testing, but in the meantime I'll just run https://keras.io/guides/keras_cv/object_detection_keras_cv/ for 50 epochs on colab on your branch.