@james77777778 2023-04-21
Hi @LukeWood , @ianstenbit 

I have investigated the issue causing the error and identified that there is a missing function/variable in VectorizedBaseImageAugmentationLayer: `self.force_output_dense_images`

Originally, JitteredResize sets `self.force_output_dense_images=True` in order to force the augmented images to become dense tensor that can be processed by `model.fit()`.

It is safe to force the images to be dense because the output images of JitteredResize must have same shape (B, target_size[0], target_size[1] C).

To resolve this issue, we can update VectorizedBaseImageAugmentationLayer by bring back `self.force_output_dense_images` functionality.

Here is a possible solution (at VectorizedBaseImageAugmentationLayer):
Change

https://github.com/keras-team/keras-cv/blob/4ce23387eb1a27cf84e594206ecdb625f4003978/keras_cv/layers/preprocessing/vectorized_base_image_augmentation_layer.py#L293-L303

to

```python
        if isinstance(images, tf.RaggedTensor):
            inputs_for_raggeds = {"transformations": transformations, **inputs}
            print("inputs_for_raggeds", inputs_for_raggeds)
            print(
                "self._unwrap_ragged_image_call", self._unwrap_ragged_image_call
            )
            images = tf.map_fn(
                self._unwrap_ragged_image_call,
                inputs_for_raggeds,
                fn_output_signature=self.compute_ragged_image_signature(images),
            )
            if self.force_output_dense_images:  # <--- use self.force_output_dense_images to control
                images = images.to_tensor()
```

And now we can set `self.force_output_dense_images` in JitteredResize's `__init__` to force the output images to be dense.

RetinaNet training works after the above modification.

WDYT?