@james77777778 2023-04-21
The running screenshot (but I changed training set to `voc/2007` to avoid downloading)

![screenshot](https://user-images.githubusercontent.com/20734616/233524232-ca840b63-2a5e-4bb7-a034-d01884b8ca7c.PNG)