@LukeWood 2023-04-21
> The running screenshot (but I changed training set to `voc/2007` to avoid downloading)
> 
> ![screenshot](https://user-images.githubusercontent.com/20734616/233524232-ca840b63-2a5e-4bb7-a034-d01884b8ca7c.PNG)

Thanks for digging into this; I feel this looks good.  Lets try to get a good numerical score on the training run of the guide, then happy to merge.  Thanks for digging into this!  

Can we also add a basic test case asserting that JitteredResize outputs dense images?