@james77777778 2023-04-21
@LukeWood 

It might take some time to complete the training. I have only run a few iterations to confirm that the preprocessing pipeline is working. Should I rerun the training to completion?

I can send a PR to update the VectorizedBaseImageAugmentationLayer. Following that, this PR should be modified to ensure that dense images are always produced.

> Can we also add a basic test case asserting that JitteredResize outputs dense images?

We can keep JitteredResize listed in `DENSE_OUTPUT_TEST_CONFIGURATIONS` in `ragged_image_test.py`:

https://github.com/keras-team/keras-cv/blob/4ce23387eb1a27cf84e594206ecdb625f4003978/keras_cv/layers/preprocessing/ragged_image_test.py#L110-L137