@james77777778 2023-04-21
Hi @LukeWood @ianstenbit 

I have submitted a PR at #1726 to add the force functionality to VectorizedBaseImageAugmentationLayer 

In this PR, some modifications are needed after the fix.